updated server read() and write()

+++ shell/server.cpp
@@ -36,8 +36,8 @@ int main()
 
     while (true)
     {
-        char readMsg[3000] = {0};                              // allocated space for receiving from client
-        int readfd = read(acceptfd, readMsg, sizeof(readMsg)); // receive message from client
+        char readMsg[3000] = {0};                                      // allocated space for receiving from client
+        int readfd = read(acceptfd, readMsg, (size_t)sizeof(readMsg)); // receive message from client
         cout << "\nClient: " << readMsg << endl;
 
         cout << "Server shell: ";
